+++ CHANGELOG.rst
@@ -2,9 +2,23 @@
 Changelog
 =========
 
-Version 0.1
-===========
+Version 0.0.1
+=============
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Added `init` command
+- Added `add-ent` command
+- Added `make-schema` command
+- Added `make-data` command
+- Added `migrate` command
+- Added `rollback` command
+- Added `info` command
+- Added `diff` command
+- Added `pull` command
+- Added `fix` command
+
+Version 0.1.0
+=============
+
+- Added support for online schema migration
+- Enhanced `info` command to display unmigrated migration plans
+- Added `pull` command